+++ apps/indexer/.env-sample
@@ -4,3 +4,4 @@ SEPOLIA_RPC=https://eth-sepolia.g.alchemy.com/v2/xxx
 ISSUER_KEY=xxx
 ISSUER_ADDRESS=0x
 DFUSE_API_KEY=server_xxx
+ALCHEMY_API_KEY=xxx
