+++ apps/indexer/package.json
@@ -19,15 +19,15 @@
         "@repo/supabase": "workspace:*",
         "@sentry/node": "^8.19.0",
         "@supabase/supabase-js": "^2.44.4",
+        "app-contracts": "workspace:*",
+        "app-env": "workspace:*",
+        "app-lib": "workspace:*",
         "axios": "^1.7.2",
         "bn.js": "^5.2.1",
         "express": "^4.19.2",
         "lodash": "^4.17.21",
         "node-fetch": "2.6.4",
         "resend": "^3.5.0",
-        "app-contracts": "workspace:*",
-        "app-env": "workspace:*",
-        "app-lib": "workspace:*",
         "viem": "latest",
         "ws": "^8.18.0"
     },
