+++ apps/indexer/src/index.ts
@@ -1,14 +1,16 @@
 import { getErrorMessage } from 'app-lib'
-import { startAuctionIndexer } from './modules/auction'
-import { startSwapsService } from './modules/swaps'
+// import { startAuctionIndexer } from './modules/auction'
+// import { startSwapsService } from './modules/swaps'
 import { startExpress } from './routes/healthcheck'
+import { startPresaleService } from './modules/presale'
 
 async function main() {
   console.log(`Launchpad indexer starting up ...`)
   try {
     startExpress()
-    startAuctionIndexer()
-    startSwapsService()
+    startPresaleService()
+    // startAuctionIndexer()
+    // startSwapsService()
   } catch (error) {
     console.log('ERROR:' + getErrorMessage(error), JSON.stringify(error))
   }
