+++ apps/indexer/src/modules/auction/easyauction.ts
@@ -1,6 +1,11 @@
-import { getContract } from 'viem'
-import { client } from '../../lib/evm-client'
+import {getContract, http, createPublicClient} from 'viem';
 import { TestnetEasyAuction } from 'app-contracts'
+import {eosEvmTestnet} from 'app-env';
+
+  const client = createPublicClient({
+    chain: eosEvmTestnet,
+    transport: http(),
+  })
 
 export const easyAuction = getContract({
   ...TestnetEasyAuction,
