+++ apps/indexer/src/modules/presale/index.ts
@@ -2,6 +2,7 @@ import { listenToEosContributions } from './eos-contributions'
 import { listenToEvmContributions } from './evm-contributions'
 
 export function startPresaleService() {
-  listenToEosContributions()
+  console.log('starting presale service')
+  // listenToEosContributions()
   listenToEvmContributions()
 }
