+++ apps/indexer/src/modules/presale/presale-issuer.ts
@@ -1,10 +1,10 @@
-import { TestnetUSDCred } from 'app-contracts'
+import { TestnetBLPL } from 'app-contracts'
 import { eosEvmTestnet } from 'app-env'
 import { createWalletClient } from 'viem'
 import { appenv } from '~/config'
 import { Address, http } from 'viem'
 
-export async function issueTokens(to: Address, amount: bigint) {
+export async function issuePresaleTokens(to: Address, amount: bigint) {
   console.log('issueTokens', {
     args: [to, amount],
   })
@@ -17,8 +17,8 @@ export async function issueTokens(to: Address, amount: bigint) {
       account: appenv.evm.issuerAccount,
     })
     return walletClient.writeContract({
-      address: TestnetUSDCred.address,
-      abi: TestnetUSDCred.abi,
+      address: TestnetBLPL.address,
+      abi: TestnetBLPL.abi,
       functionName: 'issue',
       args: [to, amount],
     })
