+++ hardhat/erc20-token/package.json
@@ -6,7 +6,8 @@
     "deployUSDCredMainnet": " hardhat run scripts/deployUSDCred.js --network eosevm_mainnet",
     "deployMBOTSPLTestnet": " hardhat run scripts/deployMBOTSPLWithFaucet.js --network eosevm_testnet",
     "deployUSDTSepolia": " hardhat run scripts/deployUSDTWithFaucet.js --network sepolia",
-    "deployUSDTTestnet": " hardhat run scripts/deployUSDTWithFaucet.js --network eosevm_testnet"
+    "deployUSDTTestnet": " hardhat run scripts/deployUSDTWithFaucet.js --network eosevm_testnet",
+    "deployBLPL": "hardhat run scripts/deployBLPL.js --network eosevm_testnet"
   },
   "devDependencies": {
     "@nomiclabs/hardhat-ethers": "^2.0.2",
@@ -15,5 +16,8 @@
     "ethereum-waffle": "^3.4.0",
     "ethers": "^5.7.2",
     "hardhat": "^2.6.4"
+  },
+  "dependencies": {
+    "@openzeppelin/contracts": "^3.3.0"
   }
 }
