+++ packages/app-contracts/src/dev/auction/testnet-deposit-order.ts
@@ -1,4 +1,5 @@
-import { EVMContractData } from "../types";
+import { sepolia } from "viem/chains";
+import { EVMContractData } from "../../types";
 
 export const TestnetDepositOrder: EVMContractData = {
   address: "0x4faA684A1E0Cdd7cb271b5424a12A2D039624D78",
@@ -6,6 +7,7 @@ export const TestnetDepositOrder: EVMContractData = {
   chainId: 15557,
   chainType: "evm",
   chainName: "EOS EVM Tesnet",
+  chain: sepolia,
   abi: [
     {
       inputs: [
