+++ packages/app-contracts/src/dev/tokens/testnet-mbots-prelaunch.ts
@@ -1,14 +1,16 @@
-import { EVMTokenContractData } from "../types";
+import { EVMTokenContractData } from "../../types";
+import {eosEvmTestnet} from '../../../../app-env/src/chains';
 
-export const SepoliaUSDT: EVMTokenContractData = {
-  address: "0x5b148580635e8b67184bcb496741e423f2c326bf",
-  name: "USDT Token",
-  symbol: "USDT",
+export const TestnetMBOTSPL: EVMTokenContractData = {
+  address: "0x357752b66961021524b44523cD90a8B3861803E5",
+  name: "MBOTS Prelaunch Token",
+  symbol: "MBOTSPL",
   decimals: 6,
-  indexFromBlock: 5394842,
-  chainId: 11155111, // sepolia
+  indexFromBlock: 30303655,
+  chainId: 15557,
   chainType: "evm",
-  chainName: "Sepolia",
+  chainName: "EOS EVM Tesnet",
+  chain: eosEvmTestnet,
   abi: [
     {
       inputs: [
@@ -44,6 +46,25 @@ export const SepoliaUSDT: EVMTokenContractData = {
         },
       ],
       name: "Approval",
+      type: "event",
+    },
+    {
+      anonymous: false,
+      inputs: [
+        {
+          indexed: true,
+          internalType: "address",
+          name: "previousOwner",
+          type: "address",
+        },
+        {
+          indexed: true,
+          internalType: "address",
+          name: "newOwner",
+          type: "address",
+        },
+      ],
+      name: "OwnershipTransferred",
       type: "event",
     },
     {
@@ -139,6 +160,19 @@ export const SepoliaUSDT: EVMTokenContractData = {
       type: "function",
     },
     {
+      inputs: [
+        {
+          internalType: "uint256",
+          name: "amount",
+          type: "uint256",
+        },
+      ],
+      name: "burn",
+      outputs: [],
+      stateMutability: "nonpayable",
+      type: "function",
+    },
+    {
       inputs: [],
       name: "decimals",
       outputs: [
@@ -218,6 +252,24 @@ export const SepoliaUSDT: EVMTokenContractData = {
       type: "function",
     },
     {
+      inputs: [
+        {
+          internalType: "address",
+          name: "recipient",
+          type: "address",
+        },
+        {
+          internalType: "uint256",
+          name: "amount",
+          type: "uint256",
+        },
+      ],
+      name: "issue",
+      outputs: [],
+      stateMutability: "nonpayable",
+      type: "function",
+    },
+    {
       inputs: [],
       name: "name",
       outputs: [
@@ -228,6 +280,26 @@ export const SepoliaUSDT: EVMTokenContractData = {
         },
       ],
       stateMutability: "view",
+      type: "function",
+    },
+    {
+      inputs: [],
+      name: "owner",
+      outputs: [
+        {
+          internalType: "address",
+          name: "",
+          type: "address",
+        },
+      ],
+      stateMutability: "view",
+      type: "function",
+    },
+    {
+      inputs: [],
+      name: "renounceOwnership",
+      outputs: [],
+      stateMutability: "nonpayable",
       type: "function",
     },
     {
@@ -309,5 +381,18 @@ export const SepoliaUSDT: EVMTokenContractData = {
       stateMutability: "nonpayable",
       type: "function",
     },
-  ] as const,
-} as const;
+    {
+      inputs: [
+        {
+          internalType: "address",
+          name: "newOwner",
+          type: "address",
+        },
+      ],
+      name: "transferOwnership",
+      outputs: [],
+      stateMutability: "nonpayable",
+      type: "function",
+    },
+  ],
+};
