+++ packages/app-contracts/src/dev/tokens/testnet-usd-cred.ts
@@ -1,14 +1,16 @@
-import { EVMTokenContractData } from "../types";
+import { EVMTokenContractData } from "../../types";
+import {eosEvmTestnet} from '../../../../app-env/src/chains';
 
-export const TestnetUSDT: EVMTokenContractData = {
-  address: "0xE561021FCB5FFB86b439ae9e6AeCE0370e2394eC",
-  name: "USDT (testing)",
-  symbol: "USDT",
+export const TestnetUSDCred: EVMTokenContractData = {
+  address: "0x1d5A4C37e60cAd0C72c057E3c191352429cDB38e",
+  name: "USD Credit Token",
+  symbol: "USDCred",
   decimals: 6,
-  indexFromBlock: 30051449,
-  chainId: 15557, // eos_evm
+  indexFromBlock: 30294882,
+  chainId: 15557,
   chainType: "evm",
   chainName: "EOS EVM Tesnet",
+  chain: eosEvmTestnet,
   abi: [
     {
       inputs: [
@@ -44,6 +46,25 @@ export const TestnetUSDT: EVMTokenContractData = {
         },
       ],
       name: "Approval",
+      type: "event",
+    },
+    {
+      anonymous: false,
+      inputs: [
+        {
+          indexed: true,
+          internalType: "address",
+          name: "previousOwner",
+          type: "address",
+        },
+        {
+          indexed: true,
+          internalType: "address",
+          name: "newOwner",
+          type: "address",
+        },
+      ],
+      name: "OwnershipTransferred",
       type: "event",
     },
     {
@@ -139,6 +160,19 @@ export const TestnetUSDT: EVMTokenContractData = {
       type: "function",
     },
     {
+      inputs: [
+        {
+          internalType: "uint256",
+          name: "amount",
+          type: "uint256",
+        },
+      ],
+      name: "burn",
+      outputs: [],
+      stateMutability: "nonpayable",
+      type: "function",
+    },
+    {
       inputs: [],
       name: "decimals",
       outputs: [
@@ -218,6 +252,24 @@ export const TestnetUSDT: EVMTokenContractData = {
       type: "function",
     },
     {
+      inputs: [
+        {
+          internalType: "address",
+          name: "recipient",
+          type: "address",
+        },
+        {
+          internalType: "uint256",
+          name: "amount",
+          type: "uint256",
+        },
+      ],
+      name: "issue",
+      outputs: [],
+      stateMutability: "nonpayable",
+      type: "function",
+    },
+    {
       inputs: [],
       name: "name",
       outputs: [
@@ -228,6 +280,26 @@ export const TestnetUSDT: EVMTokenContractData = {
         },
       ],
       stateMutability: "view",
+      type: "function",
+    },
+    {
+      inputs: [],
+      name: "owner",
+      outputs: [
+        {
+          internalType: "address",
+          name: "",
+          type: "address",
+        },
+      ],
+      stateMutability: "view",
+      type: "function",
+    },
+    {
+      inputs: [],
+      name: "renounceOwnership",
+      outputs: [],
+      stateMutability: "nonpayable",
       type: "function",
     },
     {
@@ -306,6 +378,19 @@ export const TestnetUSDT: EVMTokenContractData = {
           type: "bool",
         },
       ],
+      stateMutability: "nonpayable",
+      type: "function",
+    },
+    {
+      inputs: [
+        {
+          internalType: "address",
+          name: "newOwner",
+          type: "address",
+        },
+      ],
+      name: "transferOwnership",
+      outputs: [],
       stateMutability: "nonpayable",
       type: "function",
     },
