+++ packages/app-contracts/src/prod/tokens/usdt.ts
@@ -1,0 +1,212 @@
+import { EVMTokenContractData } from '../../types'
+import { erc20Abi } from 'abitype/abis'
+import { arbitrum, avalanche, bsc, fantom, gnosis, polygon } from 'viem/chains'
+
+export const usdtContracts: EVMTokenContractData[] = [
+  // TODO:     'Double check you have provided the correct parameters.',
+  // {
+  //   address: "0xdAC17F958D2ee523a2206206994597C13D831ec7",
+  //   name: "USDT",
+  //   symbol: "USDT",
+  //   decimals: 6,
+  //   indexFromBlock: 20432827,
+  //   chainId: 1, // Ethereum Mainnet
+  //   chainType: "evm",
+  //   chainName: "Ethereum",
+  //   abi: erc20Abi,
+  //   chain: mainnet,
+  // },
+  {
+    address: "0xFD086BC7CD5C481DCC9C85E0C3AC0A4A5404C120",
+    name: "USDT",
+    symbol: "USDT",
+    decimals: 6,
+    indexFromBlock: 1045000,
+    chainId: 42161, // Arbitrum Mainnet
+    chainType: "evm",
+    chainName: "Arbitrum",
+    abi: erc20Abi,
+    chain: arbitrum,
+  },
+  {
+    address: "0xc7198437980c041c805A1EDcbA50c1Ce5db95118",
+    name: "USDT",
+    symbol: "USDT",
+    decimals: 6,
+    indexFromBlock: 48845519,
+    chainId: 43114, // Avalanche Mainnet
+    chainType: "evm",
+    chainName: "Avalanche",
+    abi: erc20Abi,
+    chain: avalanche,
+  },
+  {
+    address: "0x55d398326f99059fF775485246999027B3197955",
+    name: "USDT",
+    symbol: "USDT",
+    decimals: 6,
+    indexFromBlock: 41076632,
+    chainId: 56, // Binance Smart Chain Mainnet
+    chainType: "evm",
+    chainName: "Binance Smart Chain",
+    abi: erc20Abi,
+    chain: bsc,
+  },
+  {
+    address: "0x049d68029688eAbF473097a2fC38ef61633A3C7A",
+    name: "USDT",
+    symbol: "USDT",
+    decimals: 6,
+    indexFromBlock: 87799484,
+    chainId: 250, // Fantom Mainnet
+    chainType: "evm",
+    chainName: "Fantom",
+    abi: erc20Abi,
+    chain: fantom,
+  },
+  {
+    address: "0x3813e82e6f7098b9583FC0F33a962D02018B6803",
+    name: "USDT",
+    symbol: "USDT",
+    decimals: 6,
+    indexFromBlock: 5000000,
+    chainId: 137, // Polygon Mainnet
+    chainType: "evm",
+    chainName: "Polygon",
+    abi: erc20Abi,
+    chain: polygon,
+  },
+  // TODO: block range
+  // {
+  //   address: "0x7F5c764cBc14f9669B88837ca1490cCa17c31607",
+  //   name: "USDT",
+  //   symbol: "USDT",
+  //   decimals: 6,
+  //   indexFromBlock: 1500000,
+  //   chainId: 10, // Optimism Mainnet
+  //   chainType: "evm",
+  //   chainName: "Optimism",
+  //   abi: erc20Abi,
+  //   chain: optimism,
+  // },
+  // {
+  //   address: "0x224E64ec1BDce3870a6a6c777eDd450454068FEC",
+  //   name: "USDT",
+  //   symbol: "USDT",
+  //   decimals: 6,
+  //   indexFromBlock: 1200000,
+  //   chainId: 1666600000, // Harmony Mainnet
+  //   chainType: "evm",
+  //   chainName: "Harmony",
+  //   abi: erc20Abi,
+  //   chain: harmonyOne,
+  // },
+  // {
+  //   address: "0x8e76f63A2dC3F3bE6f00e3F4a1ddC133D52f4866",
+  //   name: "USDT",
+  //   symbol: "USDT",
+  //   decimals: 6,
+  //   indexFromBlock: 1300000,
+  //   chainId: 1284, // Moonbeam Mainnet
+  //   chainType: "evm",
+  //   chainName: "Moonbeam",
+  //   abi: erc20Abi,
+  //   chain: moonbeam,
+  // },
+  // {
+  //   address: "0x6175F5a135dF205eD2E6FCab46dEc8df3feBf60a",
+  //   name: "USDT",
+  //   symbol: "USDT",
+  //   decimals: 6,
+  //   indexFromBlock: 1200000,
+  //   chainId: 25, // Cronos Mainnet
+  //   chainType: "evm",
+  //   chainName: "Cronos",
+  //   abi: erc20Abi,
+  //   chain: cronos,
+  // },
+  // {
+  //   address: "0x33bD5cD41aE43Fd478C14d0FF16B23E92cb9d015",
+  //   name: "USDT",
+  //   symbol: "USDT",
+  //   decimals: 6,
+  //   indexFromBlock: 1500000,
+  //   chainId: 2222, // Kava Mainnet
+  //   chainType: "evm",
+  //   chainName: "Kava",
+  //   abi: erc20Abi,
+  //   chain: kava,
+  // },
+  // {
+  //   address: "0x29db9dAF5b1D73f5A3265B5Edb2Dd9d2b99D6A3A",
+  //   name: "USDT",
+  //   symbol: "USDT",
+  //   decimals: 6,
+  //   indexFromBlock: 1200000,
+  //   chainId: 1088, // Metis Mainnet
+  //   chainType: "evm",
+  //   chainName: "Metis",
+  //   abi: erc20Abi,
+  //   chain: metis,
+  // },
+  {
+    address: "0x4ecb6e9aEa4C7D2b16B65F6375A7b88d14524eC9",
+    name: "USDT",
+    symbol: "USDT",
+    decimals: 6,
+    indexFromBlock: 1500000,
+    chainId: 100, // Gnosis Mainnet
+    chainType: "evm",
+    chainName: "Gnosis",
+    abi: erc20Abi,
+    chain: gnosis,
+  },
+  // {
+  //   address: "0x4988a896b1227218e4A686fdE5EabdcAbd91571f",
+  //   name: "USDT",
+  //   symbol: "USDT",
+  //   decimals: 6,
+  //   indexFromBlock: 1200000,
+  //   chainId: 1313161554, // Aurora Mainnet
+  //   chainType: "evm",
+  //   chainName: "Aurora",
+  //   abi: erc20Abi,
+  //   chain: aurora,
+  // },
+  // {
+  //   address: "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
+  //   name: "USDT",
+  //   symbol: "USDT",
+  //   decimals: 6,
+  //   indexFromBlock: 1000000,
+  //   chainId: 8453, // Base Mainnet
+  //   chainType: "evm",
+  //   chainName: "Base",
+  //   abi: erc20Abi,
+  //   chain: base,
+  // },
+  // {
+  //   address: "0xcebA9300f2b948710d2653dD7B07f33A8B32118C",
+  //   name: "USDT",
+  //   symbol: "USDT",
+  //   decimals: 6,
+  //   indexFromBlock: 1000000,
+  //   chainId: 42220, // Celo Mainnet
+  //   chainType: "evm",
+  //   chainName: "Celo",
+  //   abi: erc20Abi,
+  //   chain: celo,
+  // },
+  // {
+  //   address: "0x493257fD37EDB34451f62EDf8D2a0C418852bA4C",
+  //   name: "USDT",
+  //   symbol: "USDT",
+  //   decimals: 6,
+  //   indexFromBlock: 1000000,
+  //   chainId: 324, // ZKSync Era Mainnet
+  //   chainType: "evm",
+  //   chainName: "ZKSync",
+  //   abi: erc20Abi,
+  //   chain: zkSync,
+  // },
+]
